requirements.txt
absl-py==2.3.1
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.11.0
astunparse==1.6.3
attrs==25.4.0
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
decorator==4.4.2
facebook-sdk==3.1.0
Flask==3.1.2
flatbuffers==25.9.23
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==3.11
imageio==2.37.0
imageio-ffmpeg==0.6.0
instagram-private-api==1.6.0.0
instagrapi==2.2.1
instaloader==4.15
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==3.12.0
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.45.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.3
MouseInfo==0.1.3
moviepy==1.0.3
mpmath==1.3.0
namex==0.1.0
narwhals==2.10.1
networkx==3.5
numba==0.62.1
numpy==2.2.6
onnxruntime==1.23.2
opencv-python-headless==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.3
pefile==2023.2.7
pillow==12.0.0
platformdirs==4.5.0
pooch==1.8.2
proglog==0.1.12
protobuf==6.33.0
pyarrow==21.0.0
PyAutoGUI==0.9.54
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
PyGetWindow==0.0.9
Pygments==2.19.2
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
PyMatting==1.1.14
PyMsgBox==2.0.1
pyperclip==1.11.0
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-telegram-bot==22.5
pytweening==1.2.0
pytz==2025.2
pywhatkit==5.4
pywin32==311
pywin32-ctypes==0.2.3
referencing==0.37.0
rembg==2.0.67
requests==2.32.4
rich==14.2.0
rpds-py==0.28.0
scikit-image==0.25.2
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
streamlit==1.51.0
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow-hub==0.16.1
termcolor==3.2.0
tf_keras==2.20.1
tifffile==2025.10.16
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
Werkzeug==3.1.3
wheel==0.45.1
wikipedia==1.4.0
wrapt==2.0.0
yt-dlp==2025.11.12
